refactor(vm_next): isolate direct FFI adapters
